+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+### Added
+
+- optional plugin identifier
+
+### Changed
+
+- plugin discovery no for multiple base moduls of a prefix
 
 ## [0.0.2] 2022-02-25